Rename HelperFunctions.py
